Changing just the photo's title
